requirements.txt
python-telegram-bot==20.7
py-clob-client>=0.16.0
httpx==0.27.0
sqlalchemy[asyncio]==2.0.30
asyncpg==0.29.0
alembic==1.13.1
apscheduler==3.10.4
pydantic==2.7.1
pydantic-settings==2.3.0
python-dotenv==1.0.1
